feat(auth): refatorar DTOs de resposta de login e usuário
