+++ nexus-backend/src/modules/auth/auth.service.ts
@@ -4,7 +4,8 @@ import { ConfigService } from '@nestjs/config';
 import * as bcrypt from 'bcrypt';
 import { UsersService } from '../users/users.service';
 import { LoginDto } from './dto/login.dto';
-import { LoginResponseDto, UserResponseDto } from './dto/login-response.dto';
+import { LoginResponseDto } from './dto/login-response.dto';
+import { UserResponseDto } from './dto/user-response.dto';
 import { JwtPayload } from './interfaces/jwt-payload.interface';
 import { User } from '../users/entities/user.entity';
 import { AuditLogService, AuditAction, AuditCategory } from '../audit';
